requirements.txt
Flask==3.0.0
PyQt6>=6.5.0
pillow>=10.0.0
opencv-python>=4.8.0
